add warehouse operation api and fix order operation

+++ Backend/src/main/java/com/example/supermarket/dto/response/order/OrderDetailResponse.java
@@ -35,7 +35,7 @@ public class OrderDetailResponse {
     private String status;
 
     @JsonProperty("total_money")
-    private BigDecimal totalMoney;
+    private Double totalMoney;
 
     @JsonProperty("points_earned")
     private Integer pointsEarned;
